[NETTOYAGE] Déplace `afficheMagasin` dans `app/situationInventaire.js`
